requirements.txt
aiogram
mysql-connector-python
google-generativeai